remove "type": "module" and add 'serve' npm script
"type": "module" is no longer suitable for this package because it is
used for ES Module support by Node.js from v12.4. Some JS files such
as karma.conf.js are still commonjs. And current package is actually not
intended to be imported as ES Modules on Node.js, it's on browsers. So
"type": "module" is unnecessary in this case.